Code refactor KT-3
