+++ src/Application.tsx
@@ -1,7 +1,6 @@
 import useApplication from '@steroidsjs/core/hooks/useApplication';
 import HttpComponent from '@steroidsjs/core/components/HttpComponent';
 import LocaleComponent from '@steroidsjs/core/components/LocaleComponent';
-import ClientStorageComponent from '@steroidsjs/core/components/ClientStorageComponent';
 
 import 'style/index.scss';
 
@@ -14,7 +13,6 @@ export default function Application() {
         components: {
             locale: LocaleComponent,
             http: HttpComponent,
-            clientStorage: ClientStorageComponent,
         },
         onInit: ({ui}) => {
             ui.addViews(require('./ui/bootstrap').default);
