+++ src/api/service/index.ts
@@ -1,1 +1,2 @@
 export * from './currencyService';
+export * from './storageService';
