+++ src/routes/IndexPage/views/CurrencyConversionForm/views/CurrencyConversion/CurrencyConversion.scss
@@ -2,7 +2,7 @@
     width: 100%;
     display: flex;
     justify-content: space-between;
-    gap: 50px;
+    gap: 20px;
 
     &__select {
         min-width: 250px;
