+++ src/routes/IndexPage/views/CurrencyConversionForm/views/CurrencyConversion/CurrencyConversion.tsx
@@ -4,7 +4,7 @@ import {useBem, useDispatch, useSelector} from '@steroidsjs/core/hooks';
 import {DropDownField, NumberField} from '@steroidsjs/core/ui/form';
 
 import {changeInput, changeSelect, IFormState} from 'store/actions/converter';
-import {selectorCurrencyList, selectorCurrencyListIsLoading} from 'store/reducers/currency';
+import {selectCurrencyList, selectCurrencyListIsLoading} from 'store/reducers/currency';
 
 interface CurrencyConversionProps {
     form: IFormState;
@@ -18,8 +18,8 @@ function CurrencyConversion(props: CurrencyConversionProps) {
     const bem = useBem('CurrencyConversion');
     const dispatch = useDispatch();
 
-    const currencyList = useSelector(selectorCurrencyList);
-    const isLoading = useSelector(selectorCurrencyListIsLoading);
+    const currencyList = useSelector(selectCurrencyList);
+    const isLoading = useSelector(selectCurrencyListIsLoading);
 
     const items = useMemo(
         () =>
