+++ src/routes/IndexPage/views/CurrencyTable/CurrencyTable.tsx
@@ -3,11 +3,11 @@ import './CurrencyTable.scss';
 import Grid from '@steroidsjs/core/ui/list/Grid';
 import {useBem, useDispatch, useSelector} from '@steroidsjs/core/hooks';
 import getCurrencyList from 'store/actions/currency';
-import {selectorCurrencyList, selectorCurrencyListIsLoading} from 'store/reducers/currency';
+import {selectCurrencyList, selectCurrencyListIsLoading} from 'store/reducers/currency';
 import {ListQueryParams} from 'core/models';
 import getCurrencyTable from './utils';
 
-const DEFAULT_QUERY_PARAMS: ListQueryParams = {
+export const DEFAULT_QUERY_PARAMS: ListQueryParams = {
     currencies: ['RUB', 'USD', 'EUR', 'CNY'],
     numberOfCurrencies: 2,
 };
@@ -16,8 +16,8 @@ export default function CurrencyTable() {
     const bem = useBem('CurrencyTable');
     const dispatch = useDispatch();
 
-    const currencyList = useSelector(selectorCurrencyList);
-    const isLoading = useSelector(selectorCurrencyListIsLoading);
+    const currencyList = useSelector(selectCurrencyList);
+    const isLoading = useSelector(selectCurrencyListIsLoading);
 
     const table = getCurrencyTable('currencyTable', DEFAULT_QUERY_PARAMS.currencies);
