+++ src/store/reducers/converter.ts
@@ -1,7 +1,10 @@
-import {StorageService} from 'api/service/storageService';
+import {StorageService} from 'api/service';
+import {CONVERSION_FORM_ONE, CONVERSION_FORM_TWO} from 'core/constants/currencyList';
 
 import {ConverterAction, ConverterActionTypes, IFormState} from 'store/actions/converter';
 import {RootState} from '.';
+
+const DEFAULT_SELECT_ID = 1;
 
 /** Стейт формы конвертации. */
 interface ConverterState {
@@ -20,12 +23,12 @@ const initialState: ConverterState = {
     formOne: {
         inputValue: 0,
         currencyRates: undefined,
-        selectId: StorageService.get<number>('formOne'),
+        selectId: StorageService.get<number>(CONVERSION_FORM_ONE) || DEFAULT_SELECT_ID,
     },
     formTwo: {
         inputValue: 0,
         currencyRates: undefined,
-        selectId: StorageService.get<number>('formTwo'),
+        selectId: StorageService.get<number>(CONVERSION_FORM_TWO) || DEFAULT_SELECT_ID,
     },
 };
 
@@ -34,23 +37,23 @@ export const converterForm = (state = initialState, action: ConverterAction): Co
         case ConverterActionTypes.SET_FORM_ONE: {
             return {
                 ...state,
+                error: '',
                 formOne: {
                     selectId: action.payload.selectId || state.formOne.selectId,
                     inputValue: action.payload.inputValue,
                     currencyRates: action.payload.currencyRates || state.formOne.currencyRates,
                 },
-                error: '',
             };
         }
         case ConverterActionTypes.SET_FORM_TWO:
             return {
                 ...state,
+                error: '',
                 formTwo: {
                     selectId: action.payload.selectId || state.formOne.selectId,
                     inputValue: action.payload.inputValue,
                     currencyRates: action.payload.currencyRates || state.formTwo.currencyRates,
                 },
-                error: '',
             };
         case ConverterActionTypes.SET_ERROR_FORM:
             return {
@@ -62,6 +65,6 @@ export const converterForm = (state = initialState, action: ConverterAction): Co
     }
 };
 
-export const selectorFormOne = (state: RootState) => state.converter.formOne;
-export const selectorFormTwo = (state: RootState) => state.converter.formTwo;
-export const selectorError = (state: RootState) => state.converter.error;
+export const selectFormOne = (state: RootState) => state.converter.formOne;
+export const selectFormTwo = (state: RootState) => state.converter.formTwo;
+export const selectError = (state: RootState) => state.converter.error;
